feat: refactor with typescript
BREAKING CHANGE: Drop Node.js < 16 support

https://github.com/eggjs/egg/issues/5257
