+++ .eslintrc
@@ -1,3 +1,6 @@
 {
-  "extends": "eslint-config-egg"
+  "extends": [
+    "eslint-config-egg/typescript",
+    "eslint-config-egg/lib/rules/enforce-node-prefix"
+  ]
 }
