+++ .gitignore
@@ -13,3 +13,5 @@ results
 node_modules
 npm-debug.log
 coverage/
+.tshy*
+dist/
