--- .travis.yml
@@ -1,13 +1,0 @@
-sudo: false
-language: node_js
-node_js:
-  - '4'
-  - '6'
-  - '8'
-  - '10'
-install:
-  - npm i npminstall && npminstall
-script:
-  - npm run ci
-after_script:
-  - npminstall codecov && codecov
