+++ README.md
@@ -1,15 +1,14 @@
 # node-homedir
 
-Get the effective user's homedir, if node.js >= v6.0.0 .
+Get the effective user's homedir, if Node.js >= v16.0.0 .
 
-Aways return the homedir of the current executor， even execute with `sudo -u [user] node app.js` without `-i`.
+Always return the homedir of the current executor， even execute with `sudo -u [user] node app.js` without `-i`.
 
 [Reference Issue](https://github.com/nodejs/node/issues/5582)
 
 [![NPM version][npm-image]][npm-url]
 [![build status][travis-image]][travis-url]
 [![Test coverage][codecov-image]][codecov-url]
-[![David deps][david-image]][david-url]
 [![Known Vulnerabilities][snyk-image]][snyk-url]
 [![npm download][download-image]][download-url]
 
@@ -19,8 +18,6 @@ Aways return the homedir of the current executor， even execute with `sudo -u [
 [travis-url]: https://travis-ci.org/node-modules/node-homedir
 [codecov-image]: https://codecov.io/github/node-modules/node-homedir/coverage.svg?branch=master
 [codecov-url]: https://codecov.io/github/node-modules/node-homedir?branch=master
-[david-image]: https://img.shields.io/david/node-modules/node-homedir.svg?style=flat-square
-[david-url]: https://david-dm.org/node-modules/node-homedir
 [snyk-image]: https://snyk.io/test/npm/node-homedir/badge.svg?style=flat-square
 [snyk-url]: https://snyk.io/test/npm/node-homedir
 [download-image]: https://img.shields.io/npm/dm/node-homedir.svg?style=flat-square
@@ -29,15 +26,27 @@ Aways return the homedir of the current executor， even execute with `sudo -u [
 ## Installation
 
 ```bash
-$ npm install node-homedir --save
+npm install node-homedir
 ```
 
 ## Usage
 
+Commonjs
+
 ```js
-  const homedir = require('node-homedir');
-  console.log(homedir());
+const { homedir } = require('node-homedir');
+
+console.log(homedir());
+```
+
+ESM & TypeScript
+
+```ts
+import { homedir } from 'node-homedir';
+
+console.log(homedir());
 ```
 
 ## License
+
 [MIT](LICENSE)
