+++ package.json
@@ -2,12 +2,12 @@
   "name": "node-homedir",
   "version": "1.1.1",
   "description": "Get the effective user's homedir.",
-  "main": "index.js",
   "scripts": {
-    "test": "npm run lint && mocha -r intelli-espower-loader test/*.test.js",
-    "test-cov": "nyc mocha -r intelli-espower-loader test/*.test.js",
-    "lint": "eslint test *.js",
-    "ci": "npm run lint && npm run test-cov"
+    "test": "npm run lint -- --fix && egg-bin test",
+    "lint": "eslint src test --ext ts",
+    "ci": "npm run lint && egg-bin cov && npm run prepublishOnly",
+    "contributor": "git-contributor",
+    "prepublishOnly": "tshy && tshy-after"
   },
   "repository": {
     "type": "git",
@@ -25,21 +25,47 @@
   },
   "homepage": "https://github.com/node-modules/node-homedir#readme",
   "devDependencies": {
-    "eslint": "3",
-    "eslint-config-egg": "3",
-    "intelli-espower-loader": "^1.0.1",
-    "mm": "^2.0.0",
-    "mocha": "*",
-    "nyc": "^11.2.1",
-    "power-assert": "^1.4.1"
+    "@eggjs/tsconfig": "^1.3.3",
+    "@types/mocha": "^10.0.2",
+    "@types/node": "^20.8.2",
+    "egg-bin": "^6.5.2",
+    "eslint": "^8.50.0",
+    "eslint-config-egg": "^13.0.0",
+    "git-contributor": "^2.1.5",
+    "mm": "^3.3.0",
+    "tshy": "^1.2.2",
+    "tshy-after": "^1.0.0",
+    "typescript": "^5.2.2"
   },
   "files": [
-    "index.js"
+    "dist",
+    "src"
   ],
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=16.0.0"
   },
   "ci": {
-    "version": "4, 6, 8, 10"
-  }
+    "version": "16, 18, 20"
+  },
+  "tshy": {
+    "exports": {
+      "./package.json": "./package.json",
+      ".": "./src/index.ts"
+    }
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "import": {
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
+      }
+    }
+  },
+  "type": "module",
+  "types": "./dist/commonjs/index.d.ts"
 }
