+++ test/index.test.ts
@@ -1,16 +1,17 @@
-'use strict';
+import { strict as assert } from 'node:assert';
+import fs from 'node:fs';
+import os from 'node:os';
+import mm from 'mm';
+import homedir from '../src/index.js';
+import { homedir as homedir2 } from '../src/index.js';
 
-const homedir = require('..');
-const fs = require('fs');
-const mm = require('mm');
-const os = require('os');
-const assert = require('power-assert');
-
-describe('index.test.js', () => {
+describe('index.test.ts', () => {
   afterEach(mm.restore);
 
   it('should work', () => {
     assert.ok(fs.existsSync(homedir()));
+    assert.ok(fs.existsSync(homedir2()));
+    assert.equal(homedir(), homedir2());
   });
 
   it('should return homedir when process.env.HOME is not exist', () => {
@@ -28,7 +29,7 @@ describe('index.test.js', () => {
   it('should return homedir when os.userInfo() throw ENOENT error', () => {
     mm(os, 'userInfo', () => {
       const err = new Error('no such file or directory, uv_os_get_passwd');
-      err.code = 'ENOENT';
+      Reflect.set(err, 'code', 'ENOENT');
       throw err;
     });
