+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@eggjs/tsconfig",
+  "compilerOptions": {
+    "strict": true,
+    "noImplicitAny": true,
+    "target": "ES2022",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext"
+  }
+}
